Share Paid Claim

+++ app/src/main/java/com/teambrella/android/api/server/TeambrellaServer.java
@@ -333,7 +333,7 @@ public class TeambrellaServer {
                 teamIdString = uri.getQueryParameter(TeambrellaUris.KEY_TEAM_ID);
                 requestBody.addProperty(TeambrellaModel.ATTR_REQUEST_TEAM_ID, Integer.parseInt(teamIdString));
                 requestBody.addProperty(TeambrellaModel.ATTR_REQUEST_TO_ADDRESS, uri.getQueryParameter(TeambrellaUris.KEY_ADDRESS));
-                requestBody.addProperty(TeambrellaModel.ATTR_REQUEST_AMOUNT, Float.parseFloat(uri.getQueryParameter(TeambrellaUris.KEY_AMOUNT)));
+                requestBody.addProperty(TeambrellaModel.ATTR_REQUEST_AMOUNT, Double.parseDouble(uri.getQueryParameter(TeambrellaUris.KEY_AMOUNT)));
             }
             break;
             case TeambrellaUris.MUTE: {
